Move PollWithRetries into clients module
