+++ java/clients/src/main/java/sleeper/util/PollWithRetries.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.util;
+package sleeper.util;
 
 import java.util.function.BooleanSupplier;
