+++ java/clients/src/test/java/sleeper/utils/PollWithRetriesTest.java
@@ -13,10 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.util;
+package sleeper.utils;
 
 
 import org.junit.jupiter.api.Test;
+
+import sleeper.util.PollWithRetries;
 
 import java.util.Iterator;
 import java.util.List;
